control-esp32/requirements.txt
Flask==3.0.3
flask-mysqldb==2.0.0
gunicorn==22.0.0
paho-mqtt==1.6.1
